requirements.txt
pandas
slack-bolt
python-dotenv
jira 
openai>=1.0.0  